Add tratamento de excessão no controller no model Funcionario

+++ src/main/java/com/funcionarios/controller/FuncionarioController.java
@@ -1,6 +1,7 @@
 package com.funcionarios.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import javax.persistence.metamodel.Metamodel;
 import javax.validation.Valid;
@@ -18,6 +19,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.funcionarios.exception.FuncionarioNotFoundException;
 import com.funcionarios.model.Funcionario;
 import com.funcionarios.repository.FuncionarioRepository;
 
@@ -34,8 +36,14 @@ public class FuncionarioController {
     }
     
     @GetMapping("/get/{id}")
-    public Funcionario get(@PathVariable(value = "id") long id) {
-        return this.funcionarioRepository.findById(id);
+    public Funcionario get(@PathVariable(value = "id") long id) throws FuncionarioNotFoundException {
+        Optional<Funcionario> funcionario = getOption(id);
+        if(funcionario.isPresent()) {
+            return funcionario.get();
+        }
+        else {
+            throw new FuncionarioNotFoundException("Funcionario id {" + id + "} " + "não encotrado" );
+        }
     }
     
     @PostMapping("/add")
@@ -44,15 +52,30 @@ public class FuncionarioController {
     }
     
     @PutMapping("/update")
-    public void update(@Valid @RequestBody Funcionario funcionario) {
-        this.funcionarioRepository.save(funcionario);
+    public void update(@Valid @RequestBody Funcionario funcionario) throws FuncionarioNotFoundException {
+        Optional<Funcionario> f = getOption(funcionario.getId());
+        if(f.isPresent()) {
+            this.funcionarioRepository.save(f.get());
+        }
+        else {
+            throw new FuncionarioNotFoundException("Funcionario id {" + funcionario.getId() + "} " + "não encotrado" );
+        }
     }
     
     @DeleteMapping("delete/{id}")
-    public void delete(@PathVariable(value = "id") long id) {
-        this.funcionarioRepository.deleteById(id);        
+    public void delete(@PathVariable(value = "id") long id) throws FuncionarioNotFoundException {
+        Optional<Funcionario> funcionario = getOption(id);
+        if(funcionario.isPresent()) {
+            this.funcionarioRepository.deleteById(id); 
+        }
+        else {
+            throw new FuncionarioNotFoundException("Funcionario id {" + id + "} " + "não encotrado" );
+        }
     }
     
-    
+    private Optional<Funcionario> getOption(Long id){
+        Optional<Funcionario> funcionarioOptional = this.funcionarioRepository.findById(id);
+        return funcionarioOptional;
+    }
     
 }
